Fix many focus issues with the lightbox plugin
The lightbox plugin previously used a simplistic way of handling focus which didn't account for many scenario, causing keyboard traps or elements to be skipped altogether. The new solution take all content out of the DOM so that only the lightbox can be tabbed through. Exiting the lightbox restores the page in the way it was before.

Fixes #7821 #6619