Tabbed interface: Made active tab summary inherit content area's background colour by default.
Depends on #7857.

In standard tabbed interface implementations, the active tab's summary used to be set to white. This looked fine in themes that use a white background colour in their content areas (like WET). But in themes that use other background colours, it caused a mismatch between the summary's white background and the transparent background of its associated details element.

This commit resolves the mismatch by:
* Changing the $tablist-active-bg-color variable's default value to null.
* Adding if/else logic to ensure the active tab's summary uses the content area's background colour (unless a theme has specified a value for $tablist-active-bg-color).