Fix for clear button on date picker fields
The clear ("x") button on input fields doesn't appear on date picker inputs because the width is not big enough at 10em's. By setting the width to 11em, there's enough space for the "x" to appear (which now allows people to clear a date) and the field width is still small enough to fit inside a .col-md-2 as it did before.