Loading icon not cleared after seek event in Edge
In Edge (tested in v16 and v17), if you start playing a video and then seek, the playing event is not re-fired after seeked event, so the waiting/loading state is never cleared from the player. By listening to the seeked event and treating it the same as canplay, it clears the waiting/loading state so the loading icon is removed and doesn't obstruct the player as the video continues to play.