Form SCSS: Remove invalid selector and associated scss-lint rule
That selector's sub-selectors are designed to place a red wildcard in a ::before pseudo-element and red text in a inner strong element on required fields. Which makes sense on labels, legends, etc... but not on void input elements.

This also removes an inline scss-lint rule whose only purpose was to prevent the checkbox input's selector from causing a linting error. WET replaced scss-lint with sass-lint long ago.