Steps form: Add type to next previous buttons (#9459)
* Steps form: Add type to next previous buttons

Fix: The default WET action on pressing Enter in an input field is to trigger a click event on the first submit type button located on the form. Since the first button on a step form is often the generated "next" button on the first step and no type is specified for this button the form will open to the second step. Specifying the button type should fix this issue

This is a PATCH